Population Model Sensitivity Test

Runs an OAT type of sensitivity analysis on the parameters in the JOEMODEL. It can handle multiple HUCs at the same time. The sensitivity is reported for both the baselie and the CE scenarios

Usage

pop_model_sens_test(
  dir_in = NA,
  dir_out = NA,
  dose_file = "stressor_response_fixed_ARTR.xlsx",
  sr_wb_dat = "stressor_magnitude_unc_ARTR.xlsx",
  life_cycle_params = "life_cycles.csv",
  HUC_ID = NA,
  n_years = NA,
  MC_sims = NA,
  stressors = NA,
  last_years = NA,
  sens_perc_change = 10,
  sens_param = NA
)

Arguments

dir_in

Character path of the directory where the JOEMODEL input files live. Make sure that the directory holds: 1) stressor response Excel file; 2) stressor magnitude Excel file; 3) life cycle parameters Excel file

dir_out

character path to the directory where new files will be saved

dose_file

A character that specified the name of the Excel filename (with extension .xlsx) of the stressor Response file (e.g."stressor_response_fixed_ARTR.xlsx")

sr_wb_dat

A character that specified the name of the Excel filename (with extension .xlsx) of the stressor magnitude file (e.g."stressor_magnitude_unc_ARTR.xlsx")

life_cycle_params

A character that specified the name of the Excel filename (with extension .xlsx) that folds the life cycle parameters (e.g."life_cycles.csv")

HUC_ID

a numeric vector that holds the IDs of the HUCs to be assessed

n_years

A numeric number that specifies hoe long should the JOEMODEL be run.If not specified, it will default to 300

MC_sims

A number that specified how many Monte Carlo (MC) simulations will be run for each HUC. Each MC run will sample from the stressor's distribution and return the corresponding mean system capacity. If not specified; it will default to 5

stressors

A character vector that holds the stressors that you want to include when running the sensitivity analysis. These stressors are common to ALL HUCs

last_years

A numeric that specifies the number of years prior to the last that you consider that the population reached stability in. If not specified; it will assume that the last 20% of your years were stable (e.g. if your n_years was set to 100; then the default stable years will be the last 20 years). The OAT sensitivity analysis will only consider the population in those last years

sens_perc_change

A numeric that specifies the percent change in the model parameters that the sensitivity analysis will be considered (e.g. specify 10 if you want to test how a 10 percent change in the parameters to be tested will impact the population in the defined stable years). YOU can specify -10 if you want to see how reducing the parameter by 10 percent will impact the stable populations. The function assumes the same percent change across all parameters.If not specified, it will default to 10 (percent)

sens_param

A character vector that specifies the JOEMODEL parameters that you would like to run the OAT sensitivity analysis on. Please report the Name and not the Parameter (e.g. report "Nstage: and not"Number of life stages"). If not specified; it will assume that you want to run it on all parameters EXCEPT Nstage and any parameter whose value in the life cycle parameters file was set to zero.

Value

The function will return:

SA: A list where each element holds the median sensitivities of the JOEMODEL both for the baseline and the CE runs for a given HUC. The list will have the same number of elements as your HUCs

life_cycles_sens_analysis_HUCXXXXXXXXXX a set of .csv files. Each HUC will have its own .csv file. They will be saved in your dir_out

Examples

if (FALSE) {
library(JoeModelCE)
SA <- PopulationModel_Run_Sen(dir_in="C:/JOEMODEL", dir_out="C:/JOEMODEL/Results", dose_file = "stressor_response_fixed_ARTR.xlsx",
                               sr_wb_dat = "stressor_magnitude_unc_ARTR.xlsx", life_cycle_params = "life_cycles.csv",
         HUC_ID = c(1701010201, 1701010202), n_years = 300, MC_sims = 5, stressors = NA, last_years=20,
        sens_perc_change=10, sens_param=c("k","surv_1", "eps", "M.cv"))
        }